Fix issue reported by Dumas Walker: When importing a FidoNet echolist (NA file), the FidoNet address was forced to the first configured FTN AKA for the system, not copied from the template sub.

+++ smblib/smbadd.c
@@ -1,6 +1,6 @@
 /* Synchronet message base (SMB) high-level "add message" function */
 
-/* $Id: smbadd.c,v 1.40 2018/07/17 06:31:46 rswindell Exp $ */
+/* $Id: smbadd.c,v 1.42 2019/06/29 00:44:28 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -38,6 +38,7 @@
 #include "smblib.h"
 #include "genwrap.h"
 #include "crc32.h"
+#include "lzh.h"
 
 /****************************************************************************/
 /****************************************************************************/
@@ -92,7 +93,7 @@ int SMBCALL smb_addmsg(smb_t* smb, smbmsg_t* msg, int storage, long dupechk_hash
 					,__FUNCTION__
 					,smb_hashsourcetype(found.source)
 					,smb_hashsource(msg,found.source)
-					,found.number);
+					,(ulong)found.number);
 				retval=SMB_DUPE_MSG;
 				break;
 			}
